// client/src/lib/Advisories/types/csaf-2.1.d.ts
export type JSONSchema = "https://docs.oasis-open.org/csaf/csaf/v2.1/schema/csaf.json";
/**
 * Contains a list of acknowledgment elements associated with the whole document.
 *
 * @minItems 1
 */
export type DocumentAcknowledgments = [Acknowledgment, ...Acknowledgment[]];
// ...
export type DocumentCategory = string;
/**
 * Gives the version of the CSAF specification which the document was generated for.
 */
export type CSAFVersion = "2.1";
// ...
export type DocumentLanguage = string;
/**
 * Contains the SPDX license expression for the CSAF document.
 */
export type LicenseExpression = string;
/**
 * Holds notes associated with the whole document.
 *
 * @minItems 1
 */
export type DocumentNotes = [Note, ...Note[]];
// ...
export type DocumentReferences = [Reference, ...Reference[]];
// ...
export type SourceLanguage = string;
/**
 * This SHOULD be a canonical name for the document, and sufficiently unique to distinguish it from similar documents.
 */
export type TitleOfThisDocument = string;
// ...
export type DocumentLevelExtensions = [CSAFExtensionContent, ...CSAFExtensionContent[]];
// ...
export type ListOfBranches = [Branch, ...Branch[]];
// ...
export type ListOfFullProductNames = [FullProductName, ...FullProductName[]];
/**
 * Contains a list of product groups.
 *
 * @minItems 1
 */
export type ListOfProductGroups = [ProductGroup, ...ProductGroup[]];
// ...
export type ListOfProductPaths = [ProductPath, ...ProductPath[]];
// ...
export type Vulnerabilities = [Vulnerability, ...Vulnerability[]];
// ...
export type RootLevelExtensions = [CSAFExtensionContent, ...CSAFExtensionContent[]];

/**
 * Representation of security advisory information as a JSON document.
 */
export interface CSAFDocumentv2_1 {
  $schema: JSONSchema;
  document: DocumentLevelMetaData;
  product_tree?: ProductTree;
  vulnerabilities?: Vulnerabilities;
  x_extensions?: RootLevelExtensions;
}
/**
 * Captures the meta-data about this document describing a particular set of security advisories.
 */
export interface DocumentLevelMetaData {
  acknowledgments?: DocumentAcknowledgments;
  aggregate_severity?: AggregateSeverity;
  category: DocumentCategory;
  csaf_version: CSAFVersion;
  distribution: RulesForDocumentSharing;
  lang?: DocumentLanguage;
  license_expression?: LicenseExpression;
  notes?: DocumentNotes;
  publisher: Publisher;
  references?: DocumentReferences;
  source_lang?: SourceLanguage;
  title: TitleOfThisDocument;
  tracking: Tracking;
  x_extensions?: DocumentLevelExtensions;
}
// ...
export interface AggregateSeverity {
  namespace?: NamespaceOfAggregateSeverity;
  text: TextOfAggregateSeverity;
}
/**
 * Describe any constraints on how this document might be shared.
 */
export interface RulesForDocumentSharing {
  sharing_group?: SharingGroup;
  text?: TextualDescription;
  tlp: TrafficLightProtocolTLP;
}
// ...
export interface Publisher {
  category: CategoryOfPublisher;
  contact_details?: ContactDetails;
  issuing_authority?: IssuingAuthority;
  name: NameOfPublisher;
  namespace: NamespaceOfPublisher;
}
// ...
export interface Tracking {
  aliases?: Aliases;
  current_release_date: CurrentReleaseDate;
  generator?: DocumentGenerator;
  id: UniqueIdentifierForTheDocument;
  initial_release_date: InitialReleaseDate;
  revision_history: RevisionHistory;
  status: DocumentStatus;
  version: Version;
}
// ...
export interface CSAFExtensionContent {
  $schema: CSAFExtensionSchema;
  category: ExtensionCategory;
  content: Content;
  critical: Critical;
}
/**
 * Contains the additional information in its properties.
 */
export interface Content {
  [k: string]: unknown;
}
/**
 * Is a container for all fully qualified product names that can be referenced elsewhere in the document.
 */
export interface ProductTree {
  branches?: ListOfBranches;
  full_product_names?: ListOfFullProductNames;
  product_groups?: ListOfProductGroups;
  product_paths?: ListOfProductPaths;
}
// ...
export interface Vulnerability {
  acknowledgments?: VulnerabilityAcknowledgments;
  cve?: CVE;
  cwes?: ListOfCWEs;
  disclosure_date?: DisclosureDate;
  discovery_date?: DiscoveryDate;
  first_known_exploitation_dates?: ListOfFirstKnownExploitationDates;
  flags?: ListOfFlags;
  ids?: ListOfIDs;
  involvements?: ListOfInvolvements;
  metrics?: ListOfMetrics;
  notes?: VulnerabilityNotes;
  product_status?: ProductStatus;
  references?: VulnerabilityReferences;
  remediations?: ListOfRemediations;
  threats?: ListOfThreats;
  title?: Title;
  x_extensions?: VulnerabilityLevelExtensions;
}